requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve>=0.0.51
fastapi>=0.110
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox
langchain_ollama>=0.1.7
pydantic>=2.9